backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
sqlalchemy>=2.0.30
asyncpg>=0.31.0
psycopg2-binary>=2.9.11
alembic>=1.13.1
pydantic[email]>=2.7.1
pydantic-settings>=2.3.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.1
openai>=1.0.0
instructor>=1.15.1
faster-whisper>=1.2.1
faker>=25.0.0
rich>=13.0.0
questionary>=2.1.1
stripe>=8.0.0
fpdf2>=2.8.7
firebase-admin>=6.5.0
